feat: fully nixify CI to benefit from cacheing (#27)
Until now, we've been using canonical `cargo` commands for builds in
PRs. This switches everything to nix.

The reason we hadn't switched before was that I hadn't figured out how
to print full logs with nix. This is now achieved via
`--print-build-logs`, which in retrospect seems too easy to be hard to
figure out - not an uncommon feeling when working with nix, and not an
excuse per se, but alas, here we are.

This means that builds in PRs should now create the cache for builds on
main, such that we won't need to fully build the crate on main any
longer - just the Docker image.