feat: use UUIDs for GitLab Runners (#29)
Until now, our `GitLabRunner` used the `id` field as primary key in the
database, with the additional constraint `UNIQUE(id, url, token)`
because runner IDs are actually only unique per GitLab instance and each
token can only be used once per GitLab instance as well. There is an
obvious issue with that:

If the `id` is `PRIMARY KEY`, it is also `UNIQUE`, rendering the
constraint `UNIQUE(id, url, token)` meaningless. Furthermore, as a
result, we would never be able to store two runners with the same `id`,
even though it is perfectly fine to have two such runners as long as
they are not connected to the same GitLab instance.

This change fixes that by introducing a synthetic primary key in the
form of a UUID. The `UNIQUE(id, url, token)` is kept in place because it
is semantically correct: any GitLab instance generates a new `(id,
token)` combination for every runner.