feat: Map GitLab Runner Docker Parameters (#28)
We are currently only exposing a minimal set of adjustable parameters to
the REST API of runnrs. In an effort to prepare for advanced use-cases
this PR maps all available Docker Executor parameters of the GitLab
documentation to the glrcfg.

[GitLab Runners Docker
Section](https://archives.docs.gitlab.com/15.11/runner/configuration/advanced-configuration.html#the-runnersdocker-section)

We implement all types according to the specifications in the
documentation. For parameters whose default is explicitly specified, it
is taken over.

---------

Co-authored-by: Florian Eich <florian.eich@bmc-labs.com>