feat: ensure that 'multiple' in MaybeMultiple are actually multiple (#33)
In our current implementation of `MaybeMultiple`, the
`MaybeMultiple::Multiple` variant holds a `Vec`. That `Vec` could be
empty or only contain one element, which is semantically equivalent to
the `::None` and `::Some(T)` variants respectively.

This changes the `::Multiple` variant to hold a container that can only
contain 2 or more elements.