feat: relax length constraint on GitLab Runner tokens (#37)
GitLab [does not specify the Runner token
suffix](https://docs.gitlab.com/runner/register/), so the 20 alphanum
chars were merely a heuristic. It worked with our GitLab 16.x instance
but no longer does since the upgrade to 17.x - whether that's due to a
change on GitLab's side or because tokens just vary in length as more of
them get created has not been investigated.

This relaxes the constraint to allow between 16 and 32 alphanum chars.