feat: revert CI changes and re-add deploy step (#40)
The previously implemented changes to CI worked, but due to how GitHub
caches between jobs, did not produce the Docker container with the
correct version. Therefore, this reverts to the previous setup but adds
the deploy step, meaning that cutting a release will be manual again.

The only way to get the correct setup seems to be using the GitHub API
to do releases.