feat(defaults): connect host's Docker socket to runners by default (#49)
## 📝 Description

The `docker` command line tool needs access to the Docker daemon for
certain tasks, amongst them the ability to log into and push images to a
registry. This MR gives GitLab Runners access to the host's Docker
daemon, which enables building and pushing Docker images in the runner.

This also decreases security. For now, we use `runrs` on isolated
machines running Docker only for the purpose of having GitLab Runners,
and as long as `runrs` isn't used to operate multi-tenant instances,
this is a bearable compromise. For the future however, a different
solution must be found.

## 🔗 Related Issue(s)

None.

## 🎯 Type of Change

- [ ] 🐛 Fix (non-breaking change)
- [x] ✨ Feature (non-breaking change)
- [ ] 💥 Breaking change (change breaks user facing API)
- [ ] 📚 Documentation update
- [ ] ⚙️  Chore (refactoring, build process, CI, etc.)
- [ ] 🧪 Test addition or update
- [ ] ⏪ Revert

## ✅ Checklist

- [ ] ~Added tests for new code.~ Not applicable, existing tests (here
and in the `glrcfg` crate) cover this.
- [ ] CI is green, no warnings.
- [x] Code style visually aligns with the remainder of the codebase.
- [x] Comments are sensible and comprehensive.
- [x] Docs are up-to-date.
- [x] Commit message follows [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/).
- [x] PR description is accurate and comprehensive.

## 📖 Additional Notes (if any)

Future TODO: find a better solution.

---